requirements.txt
altgraph==0.17.4
dacite==1.8.1
esper==3.2
macholib==1.16.3
packaging==24.0
pygame-ce==2.4.1
pygbag==0.9.1
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3